requirements.txt
# Core Vector DB & Embeddings
pymilvus[model]>=2.4.0

sentence-transformers>=2.7.0
openai-clip>=1.0.1

# LLM Inference
llama-cpp-python[cuda]>=0.2.90

# Audio
openai-whisper>=20231117
ffmpeg-python>=0.2.0
pydub>=0.25.1
pyannote.audio>=3.1.0

# Document Parsing
PyMuPDF>=1.24.0
python-docx>=1.1.0
docx2txt>=0.8
pdfplumber>=0.10.0

# OCR & Image Processing
easyocr>=1.7.1
opencv-python-headless>=4.9.0.80
Pillow>=10.0.0

# Sparse Retrieval
tantivy>=0.22.0

# API & Frontend
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gradio>=4.36.0
aiofiles>=23.0.0

# Export
pandas>=2.2.0
openpyxl>=3.1.0
python-pptx>=0.6.23

# Utilities
tiktoken>=0.7.0
python-magic>=0.4.27
loguru>=0.7.0
python-dotenv>=1.0.0
torch>=2.2.0
transformers>=4.40.0
numpy>=1.26.0
scipy>=1.13.0
